refactor(channels): the kit and the mechanism become the two directories they already were
