+++ src/channels/registration.ts
